feat: export tax-free tags utils and filter component
